crosstool-ng does not like CC
When CC is set, the "Configuring GMP for host" step during a Canadian build uses $CC as its compiler instead of the build->host cross-compiler. Since CC is typically set to the native compiler, this can cause all sorts of problems. For example, when the build system is Linux 64-bit and the host system is MinGW 32-bit, the configure fails with the following error:
  checking ABI=32
  checking compiler gcc -O2 -g -pipe  -fexceptions ... yes
  checking compiler gcc -O2 -g -pipe  -fexceptions has sizeof(long)==4... no
  configure: error: could not find a working compiler, see config.log for details