Use -O1 instead of -O2 for build machine
Executables compiled for the build machine are only used for the creation of the cross-toolchain (i.e., they will not be part of the generated cross-toolchain). There is no need to spend a lot of compilation time to fully optimize their execution time. The difference in compilation time between -O1 and -O2 is significant, but the execution time often differs little: http://www.phoronix.com/scan.php?page=article&item=gcc_48_og